m_apartment_api/m_apartment_api/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.1
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-ipware==7.0.1
django-js-asset==2.2.0
django-oauth-toolkit==3.0.0
django-vnpay==0.0.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
inflection==0.5.1
jwcrypto==1.5.6
mailjet-rest==1.3.4
momo==0.1.0
msgpack==1.0.8
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-ipware==3.0.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==74.1.2
sh==2.0.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
yarl==1.13.1
